+++ apps/formio-server/src/middleware/projectOwner.js
@@ -4,11 +4,10 @@ var _ = require('lodash');
 
 module.exports = function(formioServer) {
   var formio = formioServer.formio;
-  var cache = require('../cache/cache')(formio);
   return function(req, res, next) {
     // Only allow admins to change owner for now.
     if (req.projectId && req.isAdmin && req.adminKey) {
-      cache.loadCurrentProject(req, (err, project) => {
+      formio.cache.loadCurrentProject(req, (err, project) => {
         if (err) {
           return next(err);
         }